[SOUND] [REFACTO] Move SoundEditor.tsx into its SoundEditor folder
Aligns SoundEditor with the MapEditor/SpriteEditor layout
(EditorName/EditorName.tsx). Rewrites the file's internal ./SoundEditor/*
imports to ./* and updates the GameEditor import path.

Verified: typecheck, lint, and build all pass.
